Refactor Cartridge application template
Add a README that describes created application
Rename app.path -> package_compat
Move test helper from test/helper/helper.lua to test/helper.lua
